Code refactoring (2)
