+++ src/main/kotlin/ru/team10/graphApp/view/GraphView.kt
@@ -38,6 +38,7 @@ class GraphView(graph: Graph, private val verticesView: List<VertexView> = empty
         vertices().forEach {
 
             add(it)
+            add(it.st)
         }
     }
 }
